requirements.txt
Flask
pandas
reportlab
requests
Pillow
WeasyPrint
python-dotenv
Office365-REST-Python-Client
django-environ